Builtin to calculate correct DB paths for a given root
Summary: In order to manage saved state properly, we need to be able to figure out where DBs are and where they should go. This builtin function, callable from Sandcastle and usable across repos, finds out where the DB for a given repo should live in order for the autoloader to use it.

Reviewed By: dneiter

Differential Revision: D26863245

fbshipit-source-id: 677a85eb8821c6be28dd5ea6324c88ccbfa03055