Implement version checking for 64-bit saved-state
Summary:
Following up on S224005 it appears that a saved-state version mismatch check was crucial in hiding and mitigating a bug in saved-state generation infrastructure.

Fixing the underlying bug is important, but improving the resilience by a simple version check is also important.

Note: previously 32-bit already implemented such a version check, but at the time of porting to 64-bit I (mistakenly) deprioritised porting the version check.

Reviewed By: madgen

Differential Revision: D26947199

fbshipit-source-id: 5a41df6391be51eae83777b93dd8236656214d58