Delete dead bespoke irgen code
Summary: We broke support for layout-sensitive builtin calls when we eliminated FCallBuiltin. Delete dead code associated with it. We're going to need an alternate solution to this problem, but it'll also have to apply to other core layout-sensitive helpers like Shapes::idx, so this code won't be useful. Instead, we'll have some way of marking a call bytecode as a sink in our graph.

Reviewed By: arnabde03

Differential Revision: D27039938

fbshipit-source-id: af9cab0718aadeb7dd9ab604a3d7d51966955ddd