consistently use Decl_provider.xyz_key instead of string
Summary:
In many parts of the code we use Decl_provider.fun_key or type_key or gconst_key. They are all aliased to string.

In other parts of the code we use string.

This diff changes from "string" to "key". It doesn't have any effect. (It's a prelude to a future diff, which will change them to use opaque types rather than string).

Differential Revision: D27054849

fbshipit-source-id: ed651a9c1198d7f6e1fae0da4e47602a796160bb