"type_key" instead of class/typedef/recorddef_key
Summary:
It doesn't make sense to distinguish between typedef, class and record symbols.

Consider "X::f". When the typechecker resolves this, it doesn't yet know if X is a class, a typedef or a record_def. It has to look in all three categories to find a match, since the three things share a namespace. We can only distinguish what "X" was after the time we've already fetched its value, i.e. (1) after naming, (2) where the name X was indeed bound; if X was unbound then we shouldn't and can't distinguish.

andrewjkennedy further writes: "Also I worry that we might have bugs lurking relating to "duplicate" symbols e.g. a typedef and a class with the same name."

Differential Revision: D27054141

fbshipit-source-id: 17c41226498683b5160cf551d61787b902c37fc5