Typing_env.get_enum should look for class only, not typedef
Summary:
I think this is a no-op change.

In D24893742 (https://github.com/facebook/hhvm/commit/418054c61922700487336c365be1b65b4da07a09) "avoid adding dependencies on hhi definitions", the implementation of get_class was changed from calling get_class to calling get_class_or_typedef. I think that change was an accident; both get_class and get_typedef_or_class both equally avoid adding dependencies on hhi definitions.

Behavior of get_enum prior to this diff, when it called get_typedef_or_class:
1. If the name was defined in the Types naming table (used for Typedefs, Classes, Recorddefs), and that definition is for a Typedef, then get_class_or_typedef added a dependency edge from "Classs name" and return the found typedef, and get_enum would subsequently return None
2. Otherwise if the name was defined in the Types naming-table, and that definition is for a Class, then get_class_or_typedef would add a dependency edge from "Class name" and return the found class, and get_enum would return either None (if Cls.enum_type is false) or Some.
3. Otherwise, it would add a dependency edge from "Class name" and return None.

Behavior of get_enum after this diff, when it only calls get_class
1. If the name was defined in the Types naming-table, and that definition is for a Class, then get_class will add a dependency edge from "Class name" and return the found class, and get_enum will return either None (if Cls.enum_type is false) or Some
2. Otherwise it will add a dependency edge from "Class name" and return None

The two behaviors end up being the same.

Why am I doing this? How did I spot it?
1. My ultimate goal of making them be opaque is to switch away from strings and use only 64bit hashes throughout the codebase for names. I'd like to never even store the string names at all, anywhere. The only place we need to turn a 64bit hash into a human-readable name is during error reporting, and that can be done by fetching the decl (or even the AST) and displaying it then.
2. To this end I'm trying to make "symbol-name" (class_key, typedef_key) be an opaque type. Therefore, get_class has to take a class_key, and get_fun has to take a fun_key, and they're all different.
3. I had to decide what type parameter "get_class_or_typedef" should take. It can't be a class_key, it can't be a fun_key, so for now I'm just setting it to "string" since a string can be turned into either class_key or typedef_key
4. I had to decide what type parameter "get_enum" should take. Because it called get_class_or_typedef, I was forced to have it take string parameter. That just felt wrong.

Ultimately, I think it's incorrect that class_key and typedef_key and record_key have different types. They all live in the same namespace, and hence they should be the same. Indeed they have to be represented the same in the depgraph, in case "A::x" encounters an unbound A, and we have to record an edge from non-existent A, and we can't tell which of the three it is. But that's work for a future diff.

Differential Revision: D27022252

fbshipit-source-id: f6a0871e1a8a83499c9e973326ecadaa068779fa