Fix getInputs for bytecode instructions that take a LocalRange
Summary:
The offset of the LocalRange (LAR) immediate was hardcoded to 1, but that's only
correct for MemoGet.  This diff fixes this for the other bytecode instructions
that take a LocalRange, namely MemoGetEager, MemoSet, MemoSetEager and AwaitAll.

It looks like the LAR was always at offset 1 when it was introduced 4 years ago (D4376969 (https://github.com/facebook/hhvm/commit/4dbc1414ec0cfceb477b46abf0c728606dbff3c0)),
but that has been broken for a few years as MemoGet/MemoSet were changed
and AwaitAll was introduced.

I noticed getInputs was missing some inputs for MemoSet, tracked it down to the LAR mishandling,
and then inspected all the other instructions that take a LAR immediate in the hhbc table:
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/hphp/runtime/vm/hhbc.h?commit=62e30b35efd744464d7eecddf38c5f67f7f07b46&lines=522-795

Reviewed By: jano

Differential Revision: D27032525

fbshipit-source-id: e08a39202ce4a0632dcf657f854fa2447e53ad01