Make certain builtins layout-sensitive
Summary: In order to play well with struct arrays, we need to be able to JIT specialized code for Shapes builtins - particularly for Shapes::idx. This diff is a general mechanism for JIT-ing specialized code for simple Hack builtins. It's only used if bespoke array-likes are enabled. Note that the mechanism applies equally well to profiling and optimizing translations - in profiling translations, it simply means that we treat the FCall bytecode as a layout-sensitive bytecode, rather than the ops within it.

Reviewed By: colavitam

Differential Revision: D27060337

fbshipit-source-id: 6f15870defc60832a289aa7d18a6e91ce0a64971