Case-insensitivity of records, make uniform
Summary:
I think there's a typo here. We're in the function "get_canon_name". Therefore we should have `~case_insensitive:true`. We already do for typedefs and classes/interfaces/traits. It's only recorddefs that are missing it. I believe this was a typo.

(it's not a typo that we'd have encountered since in practice, since recorddefs aren't shipped).

Reviewed By: Wilfred

Differential Revision: D27065424

fbshipit-source-id: e5eef7b7d96a4da3d1de197015fe9140be5729c1