Remove Typing_deps.RecordDef
Summary:
This diff is born of theory and code-reading, not of practice.

Theory: if you write "X::f" but X is unbound, then we have to make a dependency edge from unbound X to f. But X might be a typedef, a class/interface/trait, or a record-def. They all share the same namespace. Indeed if X was already a class, and then you define a record-def with name X, then the record-def might take priority over the class, requiring us to recheck the body that used "X::f".

All this implies that the depgraph edge from X to the call-site "X::f" must not distinguish between X being a class/interface/trait, a typedef, or a record-def.

Currently in Typing_deps.ml we define use one depgraph node "Class" for typedefs and class/interface/traits, and a different depgraph node "RecordDef" for record-defs. This will lead to wrong results, per the above discussion.

Practice: Although the codebase mostly uses "RecordDef" for record-defs, it actually also uses "Class" for record-defs in two places:
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/typing/typing_env_from_def.ml?commit=4e0a362923dd&lines=77-80
https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/hphp/hack/src/deps/typing_deps.ml?commit=8a70a0d5b93bbe8c1d4d8615f4ab0bf14c66920c&lines=795-801

I believe that record-defs have never really been tested in practice, that these two places are oversights, and that record-def should just use the "Class" node in the depgraph.

This diff gets rid of the RecordDef node in the depgraph.
1. The interesting change of this diff is in Typing_deps.mli and Typing_deps.ml where I get rid of RecordDef.
2. Everywhere, I just mechanically change anything that used to use Typing_deps.Dep.RecordDef to now use Typing_deps.Dep.Class
3. The only non-mechanical bit was in serverExtractStandalone. There were a few tweaks here. I think they're all theoretical, since we're never going to extract record-defs in practice.

Differential Revision: D27056661

fbshipit-source-id: 0ce0f5aa8f338f0243835e2ef94ade35d100d630