Heterogeneous lookup for struct keys
Summary:
The core of this diff is the place where we swap out the key to slot map for an F14FastMap with a heterogeneous lookup policy. This policy allows us to have different implementations of Hash and Equal for static and non-static keys. For safety, we require callers to be explicit about which lookup type they're doing. Of course, all keys that are actually in the map must be static.

I ended up making a few other changes to this code while I was at it:
1. Introduce "struct Field", the field descriptor. It's just a key for now, but I plan to extend it soon.
1. Allocate the fields inline with the layout, using the classic "variable size suffix array" trick.
1. Add comments and fix some minor naming nits.

Reviewed By: arnabde03

Differential Revision: D27105597

fbshipit-source-id: d584126d513c64789683393a5fefbdcae3b5c6ea