Factor Env out of HhasBody
Summary: While benchmarking the new internal compiler I discovered that peak memory usage spikes to ~41GB during compilation and then returns to a final RSS of ~27GB, indicating ~14GB of excess hackc memory. Profiling with jeprof showed that the bulk of this memory is from cloning scope while generating memoized methods. We only need `Env` to be part of `HhasBody` in `print.rs` and we only care about whether or not the code is namespaced, and whether the function/method has a class and parent class. I extract this information into `HhasBodyEnv` and store that instead. With this change peak memory is now ~28GB suggesting ~2GB of excess hackc memory usage.

Reviewed By: shayne-fletcher

Differential Revision: D27017778

fbshipit-source-id: 23a66d89ba0d42d9948c5ce62153001fd550c2e1