Fix use of protected members in traits
Summary:
Hack represents `private` and `protected` member access using a type
```
type ce_visibility =
  | Vpublic
  | Vprivate of string
  | Vprotected of string
```
where the string parameter is the name of the class in which the member is "regarded" as residing. For `private` members declared in traits, this parameter is updated when the trait is embedded in a class via the `use` declaration. But for `protected`, the parameter remains the name of the trait. This leads to unsoundness.

Whatever the semantics of `protected`, it should be the case that
```
trait TR { traitbody }
class C {
  use TR;
}
```
is equivalent to
```
class C {
  traitbody
}
```
By this principle, both `Vprivate` and `Vprotected` parameters should have their "declaration site" updated upon `use`. For "synthesized" members, the declaration site should already be correct (as it will have been set in the class or interface specified by `requires extends` or `requires implements`).

Differential Revision: D26949529

fbshipit-source-id: abb63f2e21a965d111186de688bdd4f754eaf76e