Add lower bounds to AST for type constants [1/?]
Summary:
Modify Annotated ASTs to include the lower bound for context constants
(desugared into type constants with a special flag set).

This is mostly a boilerplate change, and the important
type-checking logic will be updated later.
Splitting this out to assert that there are no breakages due to a vast number of trivial changes.

Reviewed By: vassilmladenov

Differential Revision: D27032072

fbshipit-source-id: 0edc02ca59243633546ecc0250ffbaa3757b6180