Log an array's initial state
Summary: Today, if we start with a static empty dict and immediately set a key on it, we'll only get one KeyOrder - the one after the set. That will result in problematic layout choices when we add required fields. To remedy this problem, log the initial state of a LoggingArray immediately upon construction. Don't log an operation, because a static constructor is free and a counted constructor is already tracked via ConstructStr / ConstructInt.

Reviewed By: colavitam

Differential Revision: D27139890

fbshipit-source-id: 19d28fc315882e8cce192c8e2ff8d2ed460e8071