kill primary error position bugs due to localization
Summary:
This diff forces to provide an `on_error` callback when constructing an `expand_env` (usually called `ety_env`) for localization. Why: errors produced by localization have a high probability to be located in a different file, for example if we localize a type we've taken from the heap. The callback will enforce that either the error is relocated to the current file or is dropped entirely with the new callback `Errors.ignore_errors`. See the added test primary_error_pos_bug.php for an idea of what errors we want to keep vs which ones we want to silence.

To simplify, some localization functions which take a hint or any other piece of AST take a new `ignore_error` parameter instead of `on_error`. If `ignore_error` is false, then these functions will use the new error callback `invalid_type_hint hint_pos` which will wrap the errors in a "Invalid type hint" message. These are I believe neither really detrimental nor beneficial to the user, but will help make ensure that primary error positions have type Pos.t instead of Pos_or_decl.t.

In places where the error message "invalid type hint" would not make sense, we sometimes use the new callback `leave_unchanged_default_invalid_type_hint_code` which leaves the error unchanged (does not wrap it) unless no error code is provided in which case it sets one (does that ever happen at all?). In the dif that makes Pos_or_decl.t abstract, I plan to change those to `assert_error_in_current_file` which will take the current error position and check that its decl corresponds to the current decl being checked, and will drop the error and log an event otherwise.

Most of the calls to localization functions now pass `ignore_error:true`, either because we are localizing types from a remote decl, or because that function has already been called at the top-level and we don't need the error twice. We ought to be more principled in where we call localize, to avoid calling it multiple times on the same things (`localize` is costly in perf), or to avoid not calling it at all and missing important errors (I found at least two important cases for that!).

Places where we still want errors:
* basically the first time we encounter a type hint: usually in typing_toplevel. E.g. parents, member types, constraints on parameters.
* each time we check constraints, e.g. Typing_phase.check_where_constraints.
* for all in-body type hints, e.g. with `is`, `as`, lambdas.

What could be improved: `on_error` should probably be a direct argument to localize and other functions instead of a field of `ety_env` because `ety_env` is sometimes shared amongst multiple Typing_phase function calls. In fact each of these functions could have a wrapper taking `ignore_error` and setting a default error callback appropriate to the function.

Reviewed By: andrewjkennedy

Differential Revision: D27007909

fbshipit-source-id: f1439f7818d4fd70651884871a3a5469d7d1fd9a