Return sub- and super-types on failure from coercion or subtyping
Summary:
As part of the `unsafe_cast` codemod, we need to provide additional information about type errors for consumption from HackAST. This information will be added using a new TAST node annotating expressions with type information at the point a subtyping or coercion failure occurs.

In order to do this, we need to return the expected and actual types at the point the subtyping or coercion fails. At the moment, subtype/coercion errors are captured as a side-effect in the typing environment. This diff adds the sub- and super-types at which typing or coercion failed to the return type of `sub_type`, `sub_type_with_dynamic_as_bottom` and `coerce` making these types available when TAST annotation nodes are added.

Differential Revision: D26985836

fbshipit-source-id: bdf6979164d104b34c28d9fb65acdc8ad10dcf12