Throwing getters for bespoke layouts
Summary:
Introduces throwing variants of the bespoke getters to avoid Uninit checks where possible. These are dispatched through a new opcode, BespokeGetThrow.

For each layout, a templated wrapper converts the NvGet{Int/Str} functions to variants that throw if the key is not found. This wrapper is also used for dispatch to vanilla arrays. Further, we introduce getThrow to ArrayData to dispatch in the generic case.

Codegen is also updated to use BespokeGetThrow where it is advantageous.

Differential Revision: D27070768

fbshipit-source-id: 684dcdce973915b79e1ec9f15abd2f2ad34d4992