One more case showing difference in handling type const inheritance between shallow and folded decl
Summary:
As in title - I have one more idea I want to try for those, but if that doesn't work we will probably have to HH_FIXME those as part of shallow decl rollout.

This might be the same as {D16388119 (https://github.com/facebook/hhvm/commit/bcc584add0f5cca7838c5106bdeaa4111ef7508c)}, but I think presence of `Tany` make it work a bit different, so just to be sure committing as test case for future fixes.

Reviewed By: vassilmladenov

Differential Revision: D27135292

fbshipit-source-id: dce6e74013da6cec9a284832fd77ad4d28dfd403