DAP debugger: Add seq ID to all responses
Summary:
The DAP protocol says responses should have their own seq ID when returning.
Right now we put those seq IDs in a javascript layer between the client and the
server. If we want to get rid of all the javascript we need to add it in the
server part.

Reviewed By: ebluestein

Differential Revision: D26077402

fbshipit-source-id: d15f54997da925db46fce2c3c93b3d1b4e8101c9