Clear caches and do a GC when memory cap is exceeded
Summary:
Typechecker workers process a bucket of files (max 1000, or 500, depending on settings) at a time, in sequence. After processing each file, the worker checks OCaml's heap size, and if it exceeds a cap (e.g. set at 500Mb), the worker exits early. If the cap is not exceeded then all files in the bucket are processed and then the worker exits. In either case, the process is restarted.

Exiting the process is an extreme solution. In this diff, which makes use of the long-lived workers capability (when `longlived_workers=true`), we instead try to continue, by flushing all the shared memory caches, forcing a major GC, and then checking the heap size against the cap again. If it still exceeds the cap, we move on to the next bucket, but otherwise we continue to process files.

Note that it's necessary to set `longlived_workers=true` in `hh.conf` to enable this behaviour.

Reviewed By: ljw1004

Differential Revision: D25598097

fbshipit-source-id: 6fc8e3abc52e197133ab9a1bfbc2ebdc38629f3e