Disable value profiling for unpredictable indirect jumps
Summary:
These functions are frequently devirtualized in grotesque ways due to value profiling: when GCC's FDO attempts to determine the frequency of the most common values of an expression during execution. This is an independent mechanism from speculative devirtualization, but it appears to make equally bad decisions when it comes to these methods. I have not determined the underlying reason for the poor decisions.

Disabling profiling for these dispatch methods is a ~0.23% win for vanilla arrays, and we anticipate a much larger win for cpgo.

(Note: this ignores all push blocking failures!)

Differential Revision: D27189818

fbshipit-source-id: 26a89074219296171505a49cca0a9bef093ecdea