Fix typing of functions with implicit return
Summary:
This Diff allows type-checking the following program, which was previously rejected:
```
function expectMixedReturn((function(int):mixed) $f):void { }

function testit():void {
  expectMixedReturn((int $x) ==> { echo "c"; });
}
```
Previously, Hack did not accept a function returning `void` in a context where a function was expected whose return type is a proper supertype of `void`, like `mixed`.
The actual fix is trivial: The subtyping check in `Typing.implicit_return` was the wrong way around: Rather than testing if the actual return type (e.g., `void`) is a subtype of the expected return type (e.g., `mixed`), it tested the reverse direction.

However, an unintentional side-effect of the existing bug was that it rejected the following two kinds of functions:
- Functions that may terminate with and without returning a value.
- Functions always returning without a value but are  annotated with a more general type, like `mixed`.

To preserve this behavior (i.e., still reject the two scenarios above after fixing the aforementioned bug) necessitated the remaining changes in this diff:

1. I extended the TAST check in `void_return_check.ml` to also reject the two kinds of functions mentioned above. See the comment at the beginning of `void_return_check.ml` for a more precise description of what functions exactly are rejected. There are a few exceptions, mostly related to dealing with function annotated with `Tany`. This was necessary to avoid having >1000 failures in www.
2. In order to implement the TAST check mentioned above, I had to make two pieces of information available to a TAST check, that was only available during type-checking so far: Whether a function may return implicitly somewhere (without a return statement) and whether a function is part of an hhi file.
I achieved this by adding a new field named `fun_tast_info` to the typing environment type (`Typing_env_types.env`), which is saved such that it is available to TAST checks.
3. As a side-effect of me reworking `Void_return_check`, I prevented it from having runtime that is quadratic in the nesting-level of lambdas.

Reviewed By: Wilfred

Differential Revision: D22067777

fbshipit-source-id: fde3f52c6d8dd7e04b3174c7994b9ccece8046d1