Add autocomplete complete support for enum atoms
Summary:
Given
```
function f<T>(<<__Atom>> HH\MemberOf<EE, T> $m): void {}
```

This is meant trigger autocomplete on `f#` with nearly the same results as `EE::`. I just send back the constants of the enum class, ignoring the `::class` constant which is used for classname (not relevant to atoms).

The enum atom rewriting works out nicely, because when we see `f#AUTO332` the parser guarantees that we're in an `Aast.Call` node, so I can just look at `Call` nodes where the first argument is an `EnumAtom`.

Reviewed By: Wilfred

Differential Revision: D27145618

fbshipit-source-id: f7a68638c98e250f232eb804233407030d31c7b5