Improve error message for hack docblock
Summary:
I wrote this error message thinking that the most common case in which it would appear is if the user was attempting to get a docblock for a function added since the last saved state.

Context: https://fb.workplace.com/groups/hacklang.eng/permalink/5334609173254321/

As it turns out, this error message most often occurs when namespace information is missing from a docblock request.  Let's update the error message to explain that.

* We'll surround the symbol with apostrophes, to make it more clear that the word is intended to be a symbol.
* Mention that it may specifically be lacking namespace information.

Reviewed By: aweis

Differential Revision: D26983779

fbshipit-source-id: 106ae753daebbf6334258b0de8b3ce3ab11e001f