Don't cache file contents
Summary:
When looking at telemetry for a full init I noticed that the number of gets in the FileCache are exactly the same as the number of adds. This indicates we read from the file once and never refer to it again. For files that need to read a lot of other files to type check, we end up holding these file contents in worker caches taking up memory.

Switching it to not use a cache seems sensible to me.

Differential Revision: D27206050

fbshipit-source-id: 7be8aa04f315d3fb93b218426f41289cf37db65e