Store Linearizations Together
Summary:
Currently we store linearizations in the same heap under different keys, yet we always fetch both linearizations at the same time. Both linearizations are largely the same with elements appearing in different orders. This leads to taking up overall more space in our shared heap due to two things:
* When we marshall OCaml values it will look for shared structures, allowing the marshalled representation to be reduced in size.
* We can achieve better compression in our shared heap if these values are stored together

This will reduce the amount of space linearizations take up in our shared heap by more than 30%.

In order to do this I change the behavior of our linearization code. When a linearization isn't present for a particular class we will compute the member and ancestor linearizations one after the other. Given the recursive nature of the code this shouldn't cause a behavioral change. If I'm wrong about this please let me know.

There is potentially more work we can do to shrink this data given the overlapping values between the two linearizations but this is a good start.

Differential Revision: D27205190

fbshipit-source-id: 403f37904af20b75b4bc6e96fd466cd8f0275337