don't Utils.strip_ns in hh_single_type_check --linearize
Summary:
This is a small diff to change the debugging output of hh_single_type_check --linearize.

REVIEWER NOTE: please read change to hh_single_type_check.ml first.

Previously, this function would print a ns-stripped version of a classname "Foo"; now it will print the full version "\Foo".

Why? It's because I'm changing classnames to be opaque types that only support comparison. Namespace-stripping won't be a defined operation on them.

(If we want to render classnames as string, that'll be a more complicated operation -- it'll involve looking up the decl via the decl-provider and getting the name there if it has it; or, if decls are "min-decls", then it'll involve looking up the filename via the naming-table and getting the ast and parsing it to find the string name. Neither one is anything we care to do right here and now!)

Differential Revision: D27204251

fbshipit-source-id: 31694d8e16eb2e6c94f54f325b04b961d0ce4b87