typo in symbol_index_test
Summary:
I noticed this minor typo in a test, where we claim that the FileInfo.fun we're adding is a Class not a fun.

It has no effect. We're only constructing the FileInfo in this test to pass it on to the symbol index builder, which ignores the NameType. NameTypes are only used elsewhere for updating the naming table.

I only noticed this because I'm making symbol-names be opaque newtypes in a different diff, and this line wasn't working.

Differential Revision: D27209319

fbshipit-source-id: 566327b72a64f4082c24b2fca3ed4a2fbd369249