Pre-filter Member lookups with a bloom filter
Summary:
For the shallow decl architecture we build a member table lazily by walking the member linearization of a class. At each step we will need to fetch the shallow decl of a class to look up a member. These structures can be large averaging 1KB when compressed in our shared memory heap. For classes with very large linearizations it can be expensive loading all of this data from shared memory and it most cases the method will appear in only one of the classes in the linearizations.

This diff attempts to reduce the cost of this algorithm by utilizing a bloom filter. A bloom filter is constructed containing the name of all the members within a shallow class. On average these filters will be 30-50x smaller than the shallow decl while maintaining a ~1% false positive rate. Utilizing these filters we will typically avoid fetching shallow decls for a class unless we know it contains a member with that name.

We take advantage of the fact that we will typically need to check multiple bloom filters by computing the hashes once and reusing it for each bloom filter look up. This is an advantage of bloom filters versus other filters (such as cuckoo filters) that might be smaller, but require computing hashes specific to each instance of the filter.

Differential Revision: D27065046

fbshipit-source-id: 84571013a06d860ed9944c8595bc2eb0f3a03720