Fix hardware counter unit test
Summary:
This test was flaky due to the use of uninitialized memory, namely the
HardwareCounterImpl::reset_values array.  Fix that by adding a call to
HardwareCounter::Reset() to properly initialize that array.

Once the issue above was fixed, the test would often fail because of the
excessive accuracy that it was expecting for the relative error.  This
diff reduces the expected accuracy to allow the test to more consistently
pass.

Reviewed By: binliu19

Differential Revision: D27254654

fbshipit-source-id: 37863af47656482b2183967bcdbb11ba050b6977