check class constant circularity
Summary:
This is the first commit of a serie that will try to prevent
class constant to be declared in illegal circular ways. This one
prevents class constants circularity (self circularity, mutual, via extends ,...)

However it does not deal correctly with enum and enum classes where careful attention to origin of constants must be considered.

Reviewed By: madgen

Differential Revision: D26463152

fbshipit-source-id: aeb0e1f2c1142a518c25aca7f6d0b7f12f4a449c