remove naming-table-creation from rust
Summary:
We're not currently using this Rust naming-table-creation code.

Also, I want to change the way that naming tables are created, so this code is just deadweight. In particular, (1) right now I want to change the datatype of FileInfo.classes, and I want to change the schema of the table in sqlite).

I'm deleting it for now. I figure once we know the final shape, then we can write something different in Rust if we need it.

Reviewed By: bobrenjc93

Differential Revision: D27209005

fbshipit-source-id: 81ac3ecfb5ffd51df3a3a59a579bb043db65c06c