FileInfo.names isn't quite what's stored in saved-state blob
Summary:
This diff concerns how we represent FileInfo.t, the list of top-level symbols in a file.

REVIEWER NOTES: Please read FileInfo.mli first.

Background: for a given file,
1. FileInfo.t - this stores hash, (stringname+position) pairs for all top-level symbols, and optionally comments.
   * We get it out of parsing a file.
   * for people interacting with saved-state naming-table, they interact with it as a FileInfo.t
   * for IDE with modified buffers, when we look up a name, we first look it up by parsing all unmodified buffers into FileInfo.t to see if the name matches any of them either case-sensitive or case-insensitive, and failing that we fall back to the naming-table
2. FileInfo.names - this is solely the stringnames for each top-level type. The map "path->FileInfo.names" is called "FAST" in a lot of the code, though I've never learned what the acronym stands for.
3. FileInfo.saved - this is basically identical to FileInfo.t, except to save space it strips positions and comments. Internally it's expressed in terms of FileInfo.names. It's an opaque type, though, exposed to the caller as FileInfo.t.

I'm planning to move from string names to 64bit hashes. What would that mean?
1. FileInfo.t and FileInfo.saved would need to store symbol-hash and lowercase-symbol-hash for each name.
2. FileInfo.names, I believe, can get away with only the symbol-hash for each name. I'll verify that in a future diff.
3. If someone needs to get full text symbol names (e.g. to build up a search index for autoccomplete, or to resolve a symbol-hash into a text-name for an error message) then parsing will have to return some kind of "FileInfo.t_plus" which additionally includes string names

There's one wildcard here - the search service. It is given a list of all the FileInfo.t it can find. In the background while the server is idle, it ingests those: it tries to use IndexBuilder (the facts parser) to extract information from the AST, but if that fails then it falls back to using the FileInfo.t. This fallback naturally needs strings so it can build up a string index. I think I'll just remove that fallback. It shouldn't be there. The true design of the search service is that it wants to parse files to get signatures (to show up in autocomplete), and these in any case have never been available in FileInfo.t. Also I don't understand any scenario we'd have a FileInfo.t but not able to parse the file.

CONCLUSION

FileInfo.t and FileInfo.saved are basically the same thing - they'll store hashes and case-insensitive hashes. FileInfo.names has a much smaller purpose and will only store hashes.

Therefore, FileInfo.saved should not be expressed in terms of FileInfo.names. That's what this diff does.

Differential Revision: D27144914

fbshipit-source-id: fa0453467bc9c3695ef462dc4f8d03cd729fc778