Show differed naming table filenames and exit on difference
Summary: Previously the saved state verifier was only printing the count of added/removed/changed files. This diff changes its behavior to print the list of added/removed/changed filenames instead, and to return an exit code of 2 when a difference in naming table/errors/hot decls is detected- this would make the sandcastle job fail so as to provide a clearer signal.

Reviewed By: scotthovestadt

Differential Revision: D27217387

fbshipit-source-id: f2a09f9fb079a719dfd49bf98a451701f9a64cb2