Proactively deduce readonlyness of properties
Summary:
This diff makes a small correction to the ty_expr method of the readonly visitor, which deduces the readonlyness of any given expression. Specifically, when checking the readonlyness of an Obj_get expression ($x->prop), etc, when the receiver is mutable, it also checks if the property is marked readonly.

This wasn't *unsafe* before because any property that's marked readonly requires to be wrapped in a readonly expression(which is how we'll check this in HackC locally), but it does make better/more proactive error messages in cases of nested property accesses:

$x->prop1->prop2 = 4;
If prop1 is a readonly property here, instead of telling the user just to wrap it in a readonly expression (and then telling the user it was readonly after they did that), we also error saying that $x->prop1 is readonly immediately.

Reviewed By: madgen

Differential Revision: D27059657

fbshipit-source-id: 77f108450e94ba4d7636a00254749045fdc3b983