fix slow test w.r.t. default values
Summary:
Default value for function/method parameters where checked with `dynamic` as a
type, but with the enforcement status of their original type, leading to a
desynchronization and some strange behavior.

In this diff, I fix the situation by aligning the `fun_param` info to consider
the hint was `dynamic`, so it is in sync with the `Tdynamic` introduced by the
slow path check.

Reviewed By: andrewjkennedy

Differential Revision: D27289997

fbshipit-source-id: b02d1b7328c7fdd2c0af9b2c41907bf4e54cd760