Avoid linking to OCaml runtime stubs in hh_decl and hh_naming_table_builder
Summary:
The deps_rust target currently links to ocamlpool and ocamlrep_custom in order to provide an OCaml FFI. These are libraries for interfacing with the OCaml runtime, and require the binary in which they are used to link against OCaml headers. In Rust binaries which don't link against OCaml headers, we hack around that requirement by linking against ocaml_ffi_mock, which stubs out the OCaml runtime. This allows us to use the non-FFI-related parts of the library in a pure-Rust binary.

But this is a bit of a code smell, I think. We could avoid the need for the mocks if we split out the FFI-related parts of the library from the rest. This diff does so for the bits of typing_deps.rs which are used in hh_decl and hh_naming_table_builder--we only need access to the hashing functions, not the types and FFI functions associated with the 64-bit depgraph.

Reviewed By: bobrenjc93

Differential Revision: D27275756

fbshipit-source-id: 0a017056a17528a74860bf569fb88f41cc85d959